Replace framework with library
It is a library. I think I've incorrectly called it a framework because you
place your code *in* it but what we rare really just doing is passing a
function. This isn't different from the multiprocessing library... which is a
library and so we will also call this a library.